requirements.txt
loguru
numpy
Pillow
psutil
pyperclip
pywinauto
random_word
opencv-python
pyautogui